test: service 테스트
- test_s3_urls.py(리펙터링 전에 썼던 테스트 코드 파일) 삭제
